+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/WatchConnectionManager.java
@@ -227,11 +227,7 @@ public class WatchConnectionManager<T extends HasMetadata, L extends KubernetesR
             @SuppressWarnings("unchecked")
             T obj = (T) object;
             // Dirty cast - should always be valid though
-            String currentResourceVersion = resourceVersion.get();
-            String newResourceVersion = ((HasMetadata) obj).getMetadata().getResourceVersion();
-            if (currentResourceVersion == null || currentResourceVersion.compareTo(newResourceVersion) < 0) {
-              resourceVersion.compareAndSet(currentResourceVersion, newResourceVersion);
-            }
+            resourceVersion.set(((HasMetadata) obj).getMetadata().getResourceVersion());
             Watcher.Action action = Watcher.Action.valueOf(event.getType());
             watcher.eventReceived(action, obj);
           } else if (object instanceof KubernetesResourceList) {
@@ -239,11 +235,7 @@ public class WatchConnectionManager<T extends HasMetadata, L extends KubernetesR
 
             KubernetesResourceList list = (KubernetesResourceList) object;
             // Dirty cast - should always be valid though
-            String currentResourceVersion = resourceVersion.get();
-            String newResourceVersion = list.getMetadata().getResourceVersion();
-            if (currentResourceVersion == null || currentResourceVersion.compareTo(newResourceVersion) < 0) {
-              resourceVersion.compareAndSet(currentResourceVersion, newResourceVersion);
-            }
+            resourceVersion.set(list.getMetadata().getResourceVersion());
             Watcher.Action action = Watcher.Action.valueOf(event.getType());
             List<HasMetadata> items = list.getItems();
             if (items != null) {